requirements.txt
fastapi
uvicorn[standard]
pydantic
openai>=1.3.5
python-dotenv
